Dropping old date functionalities and adding minimal and commonly required date conversion format
